+++ test/sql/install.sql
@@ -25,6 +25,7 @@ select relname from pg_catalog.pg_class,
 
 -- check the emaj_version_hist content
 select verh_version from emaj.emaj_version_hist;
+select emaj.emaj_get_version();
 
 -- check history
 select hist_id, hist_function, hist_event, hist_object, hist_wording, hist_user from emaj.emaj_hist order by hist_id;
@@ -34,6 +35,8 @@ delete from emaj.emaj_hist;
 \d emaj.*
 
 -- reset function calls statistics (so the check.sql output is stable with all installation paths)
+-- wait during half a second to let the statistics collector aggregate the latest stats
+select pg_sleep(0.5);
 with reset as (select funcid, pg_stat_reset_single_function_counters(funcid) from pg_stat_user_functions
                  where (funcname like E'emaj\\_%' or funcname like E'\\_%') )
   select * from reset where funcid is null;