+++ test/sql/viewer.sql
@@ -54,6 +54,7 @@ select 'select ok' as result from (select count(*) from emaj_mySchema1.myTbl1_lo
 -----------------------------
 -- authorized functions
 -----------------------------
+select emaj.emaj_get_version();
 select * from emaj.emaj_get_current_log_table('myschema1', 'mytbl1');
 select * from emaj.emaj_verify_all();
 select emaj.emaj_get_previous_mark_group('myGroup1', current_timestamp);